Move repeated dotenv logic to libs/backend
