+++ apps/mark/backend/src/index.ts
@@ -1,41 +1,13 @@
 import { Logger, LogSource, LogEventId } from '@votingworks/logging';
-import fs from 'fs';
-import * as dotenv from 'dotenv';
-import * as dotenvExpand from 'dotenv-expand';
-import { isIntegrationTest } from '@votingworks/utils';
+import { loadEnvVarsFromDotenvFiles } from '@votingworks/backend';
 import * as server from './server';
-import { MARK_WORKSPACE, NODE_ENV, PORT } from './globals';
+import { MARK_WORKSPACE, PORT } from './globals';
 import { createWorkspace, Workspace } from './util/workspace';
 
 export type { Api } from './app';
 export * from './types';
 
-const isTestEnvironment = NODE_ENV === 'test' || isIntegrationTest();
-
-// https://github.com/bkeepers/dotenv#what-other-env-files-can-i-use
-const dotEnvPath = '.env';
-const dotenvFiles: string[] = [
-  `${dotEnvPath}.${NODE_ENV}.local`,
-  // Don't include `.env.local` for `test` environment
-  // since normally you expect tests to produce the same
-  // results for everyone
-  !isTestEnvironment ? `${dotEnvPath}.local` : '',
-  `${dotEnvPath}.${NODE_ENV}`,
-  dotEnvPath,
-  !isTestEnvironment ? `../../../${dotEnvPath}.local` : '',
-  `../../../${dotEnvPath}`,
-].filter(Boolean);
-
-// Load environment variables from .env* files. Suppress warnings using silent
-// if this file is missing. dotenv will never modify any environment variables
-// that have already been set.  Variable expansion is supported in .env files.
-// https://github.com/motdotla/dotenv
-// https://github.com/motdotla/dotenv-expand
-for (const dotenvFile of dotenvFiles) {
-  if (fs.existsSync(dotenvFile)) {
-    dotenvExpand.expand(dotenv.config({ path: dotenvFile }));
-  }
-}
+loadEnvVarsFromDotenvFiles();
 
 const logger = new Logger(LogSource.VxMarkBackend);
