+++ apps/scan/backend/src/index.ts
@@ -1,8 +1,5 @@
 import * as customScanner from '@votingworks/custom-scanner';
 import { LogEventId, Logger, LogSource } from '@votingworks/logging';
-import * as dotenv from 'dotenv';
-import * as dotenvExpand from 'dotenv-expand';
-import fs from 'fs';
 import { detectUsbDrive } from '@votingworks/usb-drive';
 import {
   InsertedSmartCardAuth,
@@ -14,7 +11,8 @@ import {
   isFeatureFlagEnabled,
   isIntegrationTest,
 } from '@votingworks/utils';
-import { NODE_ENV, SCAN_WORKSPACE } from './globals';
+import { loadEnvVarsFromDotenvFiles } from '@votingworks/backend';
+import { SCAN_WORKSPACE } from './globals';
 import * as customStateMachine from './scanners/custom/state_machine';
 import * as server from './server';
 import { createWorkspace, Workspace } from './util/workspace';
@@ -22,30 +20,7 @@ import { createWorkspace, Workspace } from './util/workspace';
 export type { Api } from './app';
 export * from './types';
 
-// https://github.com/bkeepers/dotenv#what-other-env-files-can-i-use
-const dotenvPath = '.env';
-const dotenvFiles: string[] = [
-  `${dotenvPath}.${NODE_ENV}.local`,
-  // Don't include `.env.local` for `test` environment
-  // since normally you expect tests to produce the same
-  // results for everyone
-  NODE_ENV !== 'test' ? `${dotenvPath}.local` : '',
-  `${dotenvPath}.${NODE_ENV}`,
-  dotenvPath,
-  NODE_ENV !== 'test' ? `../../../${dotenvPath}.local` : '',
-  `../../../${dotenvPath}`,
-].filter(Boolean);
-
-// Load environment variables from .env* files. Suppress warnings using silent
-// if this file is missing. dotenv will never modify any environment variables
-// that have already been set.  Variable expansion is supported in .env files.
-// https://github.com/motdotla/dotenv
-// https://github.com/motdotla/dotenv-expand
-for (const dotenvFile of dotenvFiles) {
-  if (fs.existsSync(dotenvFile)) {
-    dotenvExpand.expand(dotenv.config({ path: dotenvFile }));
-  }
-}
+loadEnvVarsFromDotenvFiles();
 
 const logger = new Logger(LogSource.VxScanBackend);
