+++ libs/backend/package.json
@@ -46,6 +46,8 @@
     "@votingworks/utils": "workspace:*",
     "buffer": "^6.0.3",
     "debug": "4.3.4",
+    "dotenv": "16.3.1",
+    "dotenv-expand": "9.0.0",
     "fs-extra": "11.1.1",
     "js-sha256": "^0.9.0",
     "jszip": "^3.9.1",
