+++ libs/backend/src/index.ts
@@ -1,9 +1,10 @@
 /* istanbul ignore file */
-export * from './election_package';
 export * from './cast_vote_records';
+export * from './dotenv';
+export * from './election_package';
 export * from './exporter';
 export * from './list_directory';
+export * from './logs';
 export * from './scan_globals';
 export * from './split';
 export * from './ui_strings';
-export * from './logs';
